+++ packages/react/CHANGELOG.md
@@ -3,6 +3,145 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.152.0-next.8](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.152.0-next.7...v2.152.0-next.8) (2022-11-22)
+
+
+### Bug Fixes
+
+* **table:** change tooltip style to carbon-like ([28dec41](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/28dec41fb5eeb3e34486efa5de346e1860881034)), closes [#3510](https://github.com/carbon-design-system/carbon-addons-iot-react/issues/3510)
+* **tableexpandrow:** add sass variable for prefix ([56a2511](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/56a2511f8766bd93919f473e2e367e3e35444c9b))
+
+
+### Features
+
+* **overflowmenu:** create IotOverflowMenu based on Carbon ([53eb6ad](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/53eb6ad76d37907e968af925fba7e0905d911a6a))
+* **tabletoobar:** add auto-positioning for tooltips ([e77786f](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/e77786fa8e198eade34b321fad6641c15f52531b))
+
+
+
+
+
+# [2.152.0-next.7](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.152.0-next.6...v2.152.0-next.7) (2022-11-18)
+
+
+### Bug Fixes
+
+* **table:** preserve search filter after resetting sorting ([df93863](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/df93863499c6289dbfc5e09ddf465be766561a70)), closes [#3491](https://github.com/carbon-design-system/carbon-addons-iot-react/issues/3491)
+
+
+
+
+
+# [2.152.0-next.6](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.152.0-next.5...v2.152.0-next.6) (2022-11-18)
+
+
+### Bug Fixes
+
+* **tablereducer:** select all action includes nested rows ([a9f01bf](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/a9f01bf64411209b3518ad624eb48a851fef3273)), closes [#3641](https://github.com/carbon-design-system/carbon-addons-iot-react/issues/3641)
+
+
+
+
+
+# [2.152.0-next.5](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.152.0-next.4...v2.152.0-next.5) (2022-11-16)
+
+
+### Features
+
+* **hierarchylist:** add prop to enable horizontal scrollbar ([6de51cd](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/6de51cd6274c248d56633a13e9920f7084392087)), closes [#3639](https://github.com/carbon-design-system/carbon-addons-iot-react/issues/3639)
+
+
+
+
+
+# [2.152.0-next.4](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.152.0-next.3...v2.152.0-next.4) (2022-11-14)
+
+
+### Features
+
+* **contentformitemtitle:** new tab for tooltip link ([f7fdca0](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/f7fdca061890aea744047905bcccd51c77c9dcc8))
+
+
+
+
+
+# [2.152.0-next.3](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.152.0-next.2...v2.152.0-next.3) (2022-11-09)
+
+
+### Bug Fixes
+
+* **table:** add left padding to pagination on small screen ([085f031](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/085f0314d031cf513cb75bf2a138965699c61b23))
+* **table:** dock left for pagination buttons ([e658e39](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/e658e39ed0858136c39a5b2f2d0f3f7843a47fe8))
+* **table:** prevent carbon pagination overflow on small screens ([b0e601e](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/b0e601ed2b06d3fac5a31e656b8afd5349c28c35))
+* **table:** remove vertical scrollbar for pagination in RTL ([a4e8cf1](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/a4e8cf1751922b313d7287ff6464bd4a40fa986f))
+
+
+
+
+
+# [2.152.0-next.2](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.152.0-next.1...v2.152.0-next.2) (2022-11-07)
+
+
+### Bug Fixes
+
+* **timepickerdropdown:** adjust regex for midnight ([99804d2](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/99804d231f16e9c8c5aba0d8b846c5b6e2e7db0b)), closes [#3630](https://github.com/carbon-design-system/carbon-addons-iot-react/issues/3630)
+
+
+
+
+
+# [2.152.0-next.1](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.151.1...v2.152.0-next.1) (2022-11-04)
+
+
+
+## [2.151.1-next.1](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.151.1-next.0...v2.151.1-next.1) (2022-11-03)
+
+
+### Bug Fixes
+
+* **table:** preserve pagination state during row edit mode ([85b4f80](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/85b4f809ba84a8639e4b252830d14d58810f6f32))
+
+
+
+## [2.151.1-next.0](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.151.0...v2.151.1-next.0) (2022-11-01)
+
+
+
+# [2.151.0-next.48](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.151.0-next.47...v2.151.0-next.48) (2022-10-31)
+
+
+
+
+
+## [2.151.1](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.151.0...v2.151.1) (2022-11-04)
+
+### Bug Fixes
+
+* **datetimepicker:** empty vaue causing invalid state ([7388e10](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/7388e1031f5102c33d0f2a3e76d9d24109ee6f3c))
+
+
+## [2.151.1-next.1](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.151.1-next.0...v2.151.1-next.1) (2022-11-03)
+
+### Bug Fixes
+
+* **table:** preserve pagination state during row edit mode ([85b4f80](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/85b4f809ba84a8639e4b252830d14d58810f6f32))
+
+
+
+
+
+## [2.151.1-next.0](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.151.0...v2.151.1-next.0) (2022-11-01)
+
+
+
+# [2.151.0-next.48](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.151.0-next.47...v2.151.0-next.48) (2022-10-31)
+
+**Note:** Version bump only for package carbon-addons-iot-react
+
+
+
+
+
 # [2.151.0](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.151.0-next.47...v2.151.0) (2022-11-01)
 # [2.151.0-next.48](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.151.0-next.47...v2.151.0-next.48) (2022-10-31)